Refactor AssemblyScanner into partial classes and focused files
Split the monolithic scanner into AssemblyScanner.cs (ctor), Scan.cs
(orchestration and relationship wiring), NodeBuilders, Reflection helpers,
IL scanning, event emission detection, event cross-reference logic, and
graph extras (sub-types, dedup, SafeGetTypes). Extracted BuildRelationships
and related helpers to clarify the scan pipeline. Removed unused
allExportedTypes variable from the original Scan implementation.
